requirements.txt
Flask==3.0.2
flask_sqlalchemy==3.1.1
img2pdf==0.5.1
mysql_connector_repackaged==0.3.1
opencv_python==4.9.0.80
Pillow==10.3.0
Werkzeug==3.0.2
